+++ kyoukai/util.py
@@ -1,12 +1,20 @@
 """
 Misc utilities for usage inside the framework.
 """
+import json
+
+import typing
 from werkzeug.wrappers import Response
 
 
+# response utilities
 def as_html(text: str, code: int = 200, headers: dict = None) -> Response:
     """
     Returns a HTML response.
+    
+    .. code-block:: python
+    
+        return as_html("<h1>Hel Na</h1>", code=403)
     
     :param text: The text to return. 
     :param code: The status code of the response.
@@ -17,6 +25,49 @@ def as_html(text: str, code: int = 200, headers: dict = None) -> Response:
         headers = {}
 
     r = Response(text, status=code, headers={"Content-Type": "text/html", **headers})
+    return r
+
+
+def as_plaintext(text: str, code: int = 200, headers: dict = None) -> Response:
+    """
+    Returns a plaintext response.
+    
+    .. code-block:: python
+    
+        return as_plaintext("hel yea", code=201)
+    
+    :param text: The text to return. 
+    :param code: The status code of the response.
+    :param headers: Any optional headers.
+    :return: A new :class:`werkzeug.wrappers.Response` representing the text.
+    """
+    if headers is None:
+        headers = {}
+
+    r = Response(text, status=code, headers={"Content-Type": "text/plain", **headers})
+    return r
+
+
+def as_json(data: typing.Union[dict, list], code: int = 200, headers: dict = None,
+            *, json_encoder: json.JSONEncoder = None) -> Response:
+    """
+    Returns a JSON response.
+    
+    .. code-block:: python
+    
+        return as_json({"response": "yes", "code": 201}, code=201)
+    
+    :param data: The data to encode.
+    :param code: The status code of the response.
+    :param headers: Any optional headers.
+    :param json_encoder: The encoder class to use to encode.
+    :return: A new :class:`werkzeug.wrappers.Response` representing the JSON.
+    """
+    if headers is None:
+        headers = {}
+
+    dumped = json.dumps(data, cls=json_encoder)
+    r = Response(dumped, status=code, headers=headers)
     return r